refactor: CardImage 컴포넌트 삭제 및 MoodboardCard 추가
- B-2 디자인 수정을 반영해 InteriorStyle 스텝에 MoodboardCard 추가, 기존 CardImage 제거
- MoodboardCard는 공컴이 아니므로 InteriorStyle 폴더 안에 선언
- 164-240 반응형 적용
