+++ src/pages/imageSetup/steps/interiorStyle/MoodboardCard.css.ts
@@ -1,0 +1,111 @@
+import { style } from '@vanilla-extract/css';
+import { recipe } from '@vanilla-extract/recipes';
+
+import { colorVars } from '@styles/tokensV2/color.css';
+import { fontVars } from '@styles/tokensV2/font.css';
+import { unitVars } from '@styles/tokensV2/unit.css';
+
+// 카드 컨테이너
+// - default: 1px secondary 테두리, 둥근 사각형
+// - selected: 1.5px strong 테두리
+// - disabled: opacity 0.5
+export const card = recipe({
+  base: {
+    aspectRatio: '164 / 240',
+    position: 'relative',
+    transition: 'transform 120ms ease',
+    border: `1px solid ${colorVars.color.border.secondary}`,
+    borderRadius: unitVars.unit.radius['600'],
+    backgroundColor: 'transparent',
+    cursor: 'pointer',
+    width: '100%',
+    overflow: 'hidden',
+    selectors: {
+      '&:active': {
+        transform: 'scale(0.95)',
+      },
+    },
+  },
+  variants: {
+    state: {
+      default: {},
+      selected: {
+        border: `1.5px solid ${colorVars.color.border.strong}`,
+      },
+      disabled: {
+        opacity: 0.5,
+        cursor: 'not-allowed',
+        selectors: {
+          '&:active': {
+            transform: 'none',
+          },
+        },
+      },
+    },
+  },
+  defaultVariants: {
+    state: 'default',
+  },
+});
+
+export const image = style({
+  display: 'block',
+  objectFit: 'cover',
+  objectPosition: 'center',
+  width: '100%',
+  height: '100%',
+});
+
+// 체크박스 — 우상단에 패딩 16px 영역 안에 위치
+export const checkbox = recipe({
+  base: {
+    position: 'absolute',
+    top: 0,
+    right: 0,
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    border: 'none',
+    background: 'transparent',
+    pointerEvents: 'none',
+    padding: unitVars.unit.gapPadding['400'],
+  },
+  variants: {
+    state: {
+      // default: 흰 stroke 빈 동그라미 20×20
+      default: {},
+      // selected: 다크 배경 24×24, 흰 숫자
+      selected: {},
+    },
+  },
+});
+
+// 체크박스 안의 동그라미
+export const circle = recipe({
+  base: {
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    transition:
+      'width 120ms ease, height 120ms ease, background-color 120ms ease',
+    borderRadius: unitVars.unit.radius.full,
+  },
+  variants: {
+    state: {
+      default: {
+        border: `1.5px solid ${colorVars.color.text.inverse}`,
+        backgroundColor: 'transparent',
+        width: '2rem',
+        height: '2rem',
+      },
+      selected: {
+        border: 'none',
+        backgroundColor: colorVars.color.fill.primary,
+        width: '2.4rem',
+        height: '2.4rem',
+        color: colorVars.color.text.inverse,
+        ...fontVars.font.body_m_14,
+      },
+    },
+  },
+});
