--- src/shared/components/card/cardImage/SkeletonCardImage.tsx
@@ -1,11 +1,0 @@
-import * as styles from './CardImage.css';
-
-const SkeletonCard = () => {
-  return (
-    <div className={styles.cardcontainer({ state: 'default' })}>
-      <div className={styles.skeletonCardImg} />
-    </div>
-  );
-};
-
-export default SkeletonCard;
