--- src/stories/CardImage.stories.tsx
@@ -1,37 +1,0 @@
-import CardImageUrl from '@assets/images/cardExImg.svg?url';
-
-import CardImage from '@components/card/cardImage/CardImage';
-
-import type { Meta, StoryObj } from '@storybook/react-vite';
-
-const meta: Meta<typeof CardImage> = {
-  title: 'shared/card/CardImage',
-  component: CardImage,
-  args: {
-    selectOrder: 1,
-    disabled: false,
-    src: CardImageUrl,
-  },
-  argTypes: {
-    selectOrder: {
-      control: 'number',
-    },
-    disabled: {
-      control: 'boolean',
-    },
-    src: {
-      control: 'text',
-    },
-  },
-};
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {};
-
-export const Disabled: Story = {
-  args: {
-    disabled: true,
-  },
-};
